Changed sample interval sleep implementation to use wall clock.
Added parameter write_timeout=0 to serial port instantiation. This should not
make any difference, but we were seeing write timeout exceptions. The write
timeout exceptions were likely the symptom of corrupted communication with the
USB to RS232 transceiver device.

Added execution of a QS after each DS command. This puts the CTD into a quiescent
sleep state, allowing sampling to proceed unaffected.

Another important change involves resending commands when we receive a CTD
sample in between sending a command and receiving the response from the
command. Prior to this change, this would cause an exception (non-fatal), now
we resend the command after recognizing that a CTD sample was received instead
of the command response.

Something to look at in a future revision is to avoid invoking the sample
received callback function in cases other than when a sample is received.
Currently we invoke the callback on receipt of any data. See comment:
"# Should we avoid calling this unless len(records) > 0 ?"